add a script that tests the model's ability to articulate its classification
for classification tasks where the model performed well (i.e., all of
them, in our case), run it through an articulation tester.
this first introduces the model to a multiple choice question where the
model must select the true relationship between the sentences and the
labels. these are shuffled to be presented in a random order. if the
answer to this question is correct, then force the model to output a
freeform response which must include a description of the actualy
classifier category.

save results to a csv.